+++ src/spicelib/devices/isrc/isrcload.c
@@ -337,7 +337,7 @@ INoi1 1 0  DC 0 TRNOISE(0n 0.5n 1 10n) : generate 1/f noise
                             double V1 = trnoise_state_get(state, ckt, n1);
                             double V2 = trnoise_state_get(state, ckt, n1+1);
 
-                            value = V1 + (V2 - V1) * (time / TS - n1);
+                            value = V1 + (V2 - V1) * (time / TS - (double)n1);
                         }
 
                         /* RTS noise */
